supplier table now obeys rls
