+++ prisma/schema.prisma
@@ -104,8 +104,8 @@ model Suppliers {
     product         String
     company         String
     notes           String?
-    Organizations   Organizations @relation(fields: [organizationsId], references: [id])
-    organizationsId String
+    Organizations   Organizations @relation(fields: [organization_id], references: [id])
+    organization_id String
 
     // @@schema("auth")
 }
