+++ src/app/dashboard/suppliers/page.tsx
@@ -8,9 +8,7 @@ import Button from "~/components/ui/Button";
 import Link from "next/link";
 
 export default function Page() {
-  const { isLoading, data, isError } = api.supplier.all.useQuery({
-    organizationEmail:""
-  });
+  const { isLoading, data, isError } = api.supplier.all.useQuery();
 
   return (
     <Layout>
